Simplify holiday impact from float to integer (0-10)
Replace decimal impact values with integers across the entire calendar
pricing system. Eliminates locale issues with decimal separators and
reduces unnecessary precision. Values like 7.5 map to 7 (preserving
the min_impact:8 threshold boundary) and 9.5 maps to 10.

Includes migration that converts existing data, updates all value
objects, services, validation, seeders, and tests.

+++ app/Domain/Calendar/Services/HolidayResolver.php
@@ -122,11 +122,11 @@ final class HolidayResolver
         return $date->next(CarbonImmutable::MONDAY);
     }
 
-    private function resolveImpact(HolidayDefinitionData $definition, CarbonImmutable $observedDate): float
+    private function resolveImpact(HolidayDefinitionData $definition, CarbonImmutable $observedDate): int
     {
         $weights = $definition->baseImpactWeights;
         $dayName = self::DAY_NAMES[$observedDate->dayOfWeek];
 
-        return $weights[$dayName] ?? $weights['default'] ?? 0.0;
+        return $weights[$dayName] ?? $weights['default'] ?? 0;
     }
 }
