+++ app/Domain/Calendar/ValueObjects/BridgeDayInfo.php
@@ -6,6 +6,6 @@ final readonly class BridgeDayInfo
 {
     public function __construct(
         public int $definitionId,
-        public float $impact,
+        public int $impact,
     ) {}
 }
